dropbox desktop sync

+++ packages/dropbox-takeover.sh
@@ -7,12 +7,16 @@ if [ ! -d ~/Dropbox/Office ]; then
 fi
 
 # Takeover Desktop
-if [ -d ~/Dropbox/Office/Desktop ]; then
-  # If exists, move contents
-  sudo mv ~/Desktop/* ~/Dropbox/Office/Desktop/
-  sudo rm -f ~/Desktop
+if [ ! -L ~/Desktop ]; then
+  if [ -d ~/Dropbox/Office/Desktop ]; then
+    # If exists, move contents
+    sudo mv ~/Desktop/* ~/Dropbox/Office/Desktop/
+    sudo rm -rf ~/Desktop
+  else
+    # If doesnt exist, create and link
+    sudo mv ~/Desktop ~/Dropbox/Office/Desktop
+  fi
+  ln -s ~/Dropbox/Office/Desktop ~/Desktop
 else
-  # If doesnt exist, create and link
-  sudo mv ~/Desktop ~/Dropbox/Office/Desktop
+  echo 'Already syncing Desktop'
 fi
-ln -s ~/Dropbox/Office/Desktop ~/Desktop
